Build - Always rebuild build.prop file
build.prop file was not rebuilt after changes to
$(PRODUCT_PROPERTY_OVERRIDES)
Without this patch, if change I681fd92e will be merged,
the wrong date will be always in the zip file's name
(date of building build.prop and not the current date).
This patch simply forces rebuilding build.prop every time.

Change-Id: I4063ddbb46069c7a4e63b354671fa939ba400009
Signed-off-by: STELIX <ssspinni@gmail.com>