Revert "core: copy installed ramdisk and bootloader to BOOT/"
This reverts commit 3f39d419f474bfbc2286b23e092709fcd44f076b.

Since this is causing problems it should be reverted. I only wrote it to get Encore booting. It will probably never become official and need this commit. We can just keep cherry-picking it for personal builds.
If other devices need it because of similarly weird boot methods it can be kept but Fugu is a lot more important than Encore.

Change-Id: I123d33c1559fb8b271a849e978c6d5666628884e